Some fixes and additions
- Added color attribute (for transparency changing) to opengl, moved
coordinate transformations to shaders
- Fixed a bug with chat bubbles
- Fixed a bug with main loop which would occur at high fps

+++ Constants.h
@@ -23,6 +23,6 @@ namespace Constants
 	// Timestep, e.g. the granularity in which the game advances.
 	const uint16_t TIMESTEP = 8;
 
-	const double VIEWWIDTH = 800.0;
-	const double VIEWHEIGHT = 600.0;
+	const float VIEWWIDTH = 800.0f;
+	const float VIEWHEIGHT = 600.0f;
 }
